qwerty/requirements.txt
Django>=5.2,<6
djangorestframework
channels
channels_redis
celery
redis
kafka-python
requests
gunicorn
daphne
drf-yasg
djangorestframework
